worker/models/module.ts: moves the model file into this subdir
